+++ controllers/routers/posts.js
@@ -16,7 +16,7 @@ const { fetch } = require('../middleware/fetch');
 // ==============================
 exports.editor.post = {
     get: (req, res, next) => {
-        req.session._view = { post: {} };
+        req.session.chest = { post: {} };
         return next();
     },
     post: async (req, res) => {
@@ -28,11 +28,11 @@ exports.editor.post = {
 
 exports.editor.edit = {
     alias: async (req, res) => {
-        req.session._view = { post: await postsModel.findOne(req.params) };
-        return res.redirect(`/posts/edit/${req.session._view.post._id}`);
+        req.session.chest = { post: await postsModel.findOne(req.params) };
+        return res.redirect(`/posts/edit/${req.session.chest.post._id}`);
     },
     get: async (req, res, next) => {
-        if (!req.session._view) req.session._view = { post: await postsModel.findById(_U_.string.readMongoId(req.url)) };
+        if (!req.session.chest) req.session.chest = { post: await postsModel.findById(_U_.string.readMongoId(req.url)) };
         return next();
     },
     patch: async (req, res) => {
@@ -49,15 +49,15 @@ exports.editor.edit = {
 
 exports.posts.show = {
     alias: async (req, res) => {
-        if (req.session.user) req.session._view = { post : await postsModel.findOne(req.params) };
-        else req.session._view = { post : await postsModel.findOne({ ...req.params, status: 'published' })};
-        return res.redirect(`/posts/${req.session._view.post.canonical}`);
+        if (req.session.user) req.session.chest = { post : await postsModel.findOne(req.params) };
+        else req.session.chest = { post : await postsModel.findOne({ ...req.params, status: 'published' })};
+        return res.redirect(`/posts/${req.session.chest.post.canonical}`);
     },
     get: async (req, res, next) => {    // tofix: post not found page
-        if (!req.session._view) {
+        if (!req.session.chest) {
             const query = { canonical: req.params[0] };
-            if (req.session.user) req.session._view = { post : await postsModel.findOne(query)};
-            else req.session._view = { post : await postsModel.findOne({ ...query, status: 'published' })};
+            if (req.session.user) req.session.chest = { post : await postsModel.findOne(query)};
+            else req.session.chest = { post : await postsModel.findOne({ ...query, status: 'published' })};
         } return next();
     },
 };